// vitals/core/src/models/cart-summary-inputs.model.ts
export interface CartSummaryInputs {
  config?: string;
  open: string;
  showCoupon: string;
  title: string;
  summaryTitle: string;
  checkoutUrl: string;
  checkoutEndpoint: string;
  continueShoppingUrl: string;
  showShipping: string;
  showTax: string;
  theme: string;
  variant: string;
  variantKey: string;
}